Moved check_ecfp_collisions to featurization to prevent circular imports
